Figured out "$@" nonsense
So cool story: I'm pretty certain that using array slices with "$@" in a
function is undocumented, or unintended. E.g., this...

> printf '[%s]\n' "${@::5}"

..includes the script's $0 as the 1st parameter, followed by any
arguments that are passed. So `foo 1 2 3` would print `filename 1 2 3`.
Very confusing.

I can not find any references to this behavior, or using "$@" as an
array (with slices, parameter expansion, n' whatnot). Seems that
substitutions behave as expected, and do not prepend $0.

Huh. Given:  foo 'this' 'that' ...

 ${@/th/ch}   no prepended $0
 ${@#th}      no prepended $0
 ${@:0}          prepended $0